requirements.txt
annotated-types==0.7.0
anyio==4.4.0
async-timeout==4.0.3
boto3==1.34.145
botocore==1.34.145
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
coverage==7.6.0
Deprecated==1.2.14
distlib==0.3.8
distro==1.9.0
exceptiongroup==1.2.2
filelock==3.15.4
fsspec==2024.6.1
gitdb==4.0.11
GitPython==3.1.43
h11==0.14.0
hatch==1.12.0
hatchling==1.25.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.24.0
hyperlink==21.0.0
idna==3.7
importlib_metadata==7.1.0
iniconfig==2.0.0
jaraco.classes==3.4.0
jaraco.context==5.3.0
jaraco.functools==4.0.2
Jinja2==3.1.4
jmespath==1.0.1
jsonpickle==3.2.2
keyring==25.3.0
libcst==1.4.0
lupa==2.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
more-itertools==10.4.0
mpmath==1.3.0
networkx==3.3
numpy==1.26.4
openai==1.36.1
opentelemetry-api==1.25.0
packaging==24.1
pathspec==0.12.1
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.2.2
pluggy==1.5.0
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
pyrsistent==0.20.0
pytest==8.3.1
pytest-asyncio==0.23.8
pytest-cov==5.0.0
python-dateutil==2.9.0.post0
PyYAML==6.0.1
redis==5.0.7
regex==2024.5.15
requests==2.32.3
rich==13.7.1
s3transfer==0.10.2
safetensors==0.4.3
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sympy==1.13.1
tokenizers==0.19.1
tomli==2.0.1
tomli_w==1.0.0
tomlkit==0.13.2
torch==2.3.1
torchaudio==2.3.1
torchvision==0.18.1
tqdm==4.66.4
transformers==4.42.4
trove-classifiers==2024.7.2
typer==0.12.3
typing_extensions==4.12.2
urllib3==2.2.2
userpath==1.9.2
uv==0.2.36
virtualenv==20.26.3
websockets==12.0
wrapt==1.16.0
zipp==3.19.2
zstandard==0.23.0
